Frontend-Backend-database-securit-rolebase_V1

+++ code/front/src/Pages/HomePage_Manager.js
@@ -13,13 +13,13 @@ export default function HomePage_Manager() {
     const[users, setUsers] = useState([])
 
     useEffect(() =>{
-        loadStaff(); 
+        loadStaff();
     },[])
 
 
     const loadStaff = async() =>{
         try {
-            const result = await api.get("/allemployee");
+            const result = await api.get("/admin/allemployee");
             setUsers(result.data);
         } catch (error) {
             console.log("error from homepage_manager");
@@ -28,7 +28,7 @@ export default function HomePage_Manager() {
 
     const deleteUser = async(id)=>{
         try {
-            await api.delete(`/${id}`);
+            await api.delete(`/admin/${id}`);
             loadStaff();
         } catch (error) {
             console.log(error);
@@ -37,55 +37,57 @@ export default function HomePage_Manager() {
 
 
     return (
-    <div className='container'>
+        <div className='container'>
 
-        <div className='py-4'>
-        <table className="table border shadow">
-  <thead>
-    <tr>
-      <th scope="col">#</th>
-      <th scope="col">ID</th>
-      <th scope="col">Name</th>
-      <th scope="col">Job Title</th>
-      <th scope="col">Annual Salary</th>
-    </tr>
-  </thead>
-  <tbody>
+            <Navbar />
 
-        {users.map((user, index) => (
-            <tr>
-            <th scope="row" key={index}> {index+1} </th>
-            <td>{user.id}</td>
-            <td>{user.name}</td>
-            <td>{user.jobTitle}</td>
-            <td>${user.annualSalary}</td>
+            <div className='py-4'>
+                <table className="table border shadow">
+                    <thead>
+                    <tr>
+                        <th scope="col">#</th>
+                        <th scope="col">ID</th>
+                        <th scope="col">Name</th>
+                        <th scope="col">Job Title</th>
+                        <th scope="col">Annual Salary</th>
+                    </tr>
+                    </thead>
+                    <tbody>
 
-            <td>
-                <Link className='btn btn-primary mx-2'
-                to={`/viewstaff/${user.id}`}>
-                    View Employee
-                </Link>
+                    {users.map((user, index) => (
+                        <tr>
+                            <th scope="row" key={index}> {index+1} </th>
+                            <td>{user.id}</td>
+                            <td>{user.name}</td>
+                            <td>{user.jobTitle}</td>
+                            <td>${user.annualSalary}</td>
 
-                <Link className='btn btn-outline-primary mx-2'
-                to={`/editstaff/${user.id}`}>
-                    Edit Employee
-                </Link>
+                            <td>
+                                <Link className='btn btn-primary mx-2'
+                                      to={`/viewstaff/${user.id}`}>
+                                    View Employee
+                                </Link>
 
-                <button className='btn btn-outline-danger mx-2'
-                onClick={()=>deleteUser(user.id)}>
-                    Delete Employee
-                </button>
-                
-            </td>
+                                <Link className='btn btn-outline-primary mx-2'
+                                      to={`/editstaff/${user.id}`}>
+                                    Edit Employee
+                                </Link>
 
-          </tr>
-        ))
-    }
+                                <button className='btn btn-outline-danger mx-2'
+                                        onClick={()=>deleteUser(user.id)}>
+                                    Delete Employee
+                                </button>
 
-  </tbody>
-</table>
+                            </td>
+
+                        </tr>
+                    ))
+                    }
+
+                    </tbody>
+                </table>
+            </div>
+
         </div>
-
-    </div>
-  )
+    )
 }
